+++ src/components/tags/Tags.js
@@ -1,4 +1,3 @@
-import { useEffect, useState } from "react"
 
 export const Tags = ({tags}) => {
